Drawing dynamic content and implement useContext and update access token
