+++ src/components/FlightResults/index.jsx
@@ -1,6 +1,7 @@
 import './styles.scss'
 
 const FlightResults = ({ children }) => {
+  
   return (
     <div className='cards-container'>
       {children}
