+++ src/components/FlightResultsCard/index.jsx
@@ -5,13 +5,13 @@ import line from '../../assets/icons/trayecto.svg'
 import departure from '../../assets/icons/departure.svg'
 import './styles.scss'
 
-const FlightResultsCard = () => {
+const FlightResultsCard = ({ data }) => {
   return (
     <article className='card cursor-pointer'>
       <section className='card__airline-info'>
         <img className='card__airline-info--icon' src={iconAirline} alt='logo airline' />
         <div className='card__airline-info--name-luggage desktop'>
-          <h4 className='name-airline'>Altitude Airways</h4>
+          <h4 className='name-airline'>{data.validatingAirlineCodes[0]}</h4>
           <div className='info'>
             <img src={luggage} alt='luggage icon' />
             <span>23kg</span>
@@ -22,7 +22,7 @@ const FlightResultsCard = () => {
         <h3 className='card__time--hour desktop'>10:25PM - 7:06AM</h3>
         <div className='card__time--duration'>
             <img src={clock} alt='clock icon' />
-            <span>10 hr 41 min</span>
+            <span>{data.itineraries[0].duration}</span>
           </div>
       </section>
       <section className='card__stops'>
@@ -33,7 +33,7 @@ const FlightResultsCard = () => {
         </div>
       </section>
       <section className='card__price'>
-        <h2 className='card__price--quantity'>275.5 USD <span>/pax</span></h2>
+        <h2 className='card__price--quantity'>{data.price.total} {data.price.currency}<span>/pax</span></h2>
         <button className='card__price--button desktop'>Choose</button>
       </section>
     </article>
