+++ src/context/FlightSearchContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const FlightSearchContext = createContext()
