+++ src/main.jsx
@@ -3,11 +3,14 @@ import { BrowserRouter } from 'react-router-dom'
 import ReactDOM from 'react-dom/client'
 import AppRoutes from './routes/AppRoutes'
 import './styles/global.scss'
+import FlightSearchProvider from './context/FlightSearchProvider'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <BrowserRouter>
-      <AppRoutes />
+      <FlightSearchProvider>
+        <AppRoutes />
+      </FlightSearchProvider>
     </BrowserRouter>
   </React.StrictMode>,
 )
